+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Added
 
 * Add support for flags: pushed lines are no longer commented.
 * Add previously used aliases as hints to new timesheets.
+* Add `--today`, `--since` and `--until` support for `status` and `commit` commands.
 
 Changed
 -------